Refactor PatientActionListener to run in Daemon thread.
